Python3.12 (#386)
* pre commit and python upgrade

* black compatible import reordering

* black compatible import reordering

* adding keras_kwargs and switchung to tuple in __get_item__

* data_2d mypy

* mypy data_2d

* mypy data3d

* mypy model_2d

* mypy test naive fusion

* np unsigend for timeseires

* mypy naive_fusion

* mypy

* mypy doesnt fail

* packaging

* data fixed

* dtype changed

* correct test python versions and ignore safety

* update tensorflow because of vulnerability

* docs building

* raising major version of actions in workflow

* fix: coverage --parallel to --parallel-mode

* upload / download artifacts to version 4

* fix: v4.4 -> v4

* fix: separate names and merge-multiple

+++ tests/test_sample_patches.py
@@ -1,4 +1,5 @@
 """Test the sample_patches."""
+
 import numpy as np
 
 from merge_stardist_masks.sample_patches import sample_patches
@@ -8,7 +9,7 @@ def test_sample_patches() -> None:
     """Run through function once and verify that shapes match."""
     shape = (3, 16, 16)
     x = np.random.random(shape)
-    y = np.zeros(shape, dtype=np.uint8)
+    y = np.zeros(shape)
     y[:, 4:8, 4:8] = 1
 
     x, y = sample_patches((y, x), patch_size=(8, 8), n_samples=1)
